+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [#1394](https://github.com/crypto-org-chain/cronos/pull/1394) Add icahost wirings but disable in parameters.
 * [#1407](https://github.com/crypto-org-chain/cronos/pull/1407) Add end-to-end encryption module.
 * [#1414](https://github.com/crypto-org-chain/cronos/pull/1414) Integrate new evm tx format.
+* [#1458](https://github.com/crypto-org-chain/cronos/pull/1458) Adjust require gas for recvPacket when ReceiverChainIsSource.
 
 ### Improvements